POC: Feature discovery for any widget
This is an attempt to add support discovery for any widget.

Instead of painting a circle with a solid color, a ring with transparent
inner part is painted. The inner part has also offset to align with the
child widget and the Ripple effect. This way overlay can be pointed to
any widget.

There are still some limittations. When TextInput is provided as a
child, transparent circle will centered, instead of showing the part
which has coursor. Also internal circle width cannot be changed.

I'm mostly conrerned if I'm going in the right direction or this is
completly flawed approach.

Fixes: #65