VisibilityDetector: Add a mechanism to force firing a callback
Add a mechanism to guarantee a visibility callback, regardless of the
widget's current or previous visibility.  Note that this will work
only for visible widgets (or for widgets that have very recently
become non-visible) since we do not want to keep track of a
potentially infinite number of previously-visible widgets.

If we do need to fire callbacks for previously-visible
VisibilityDetector widgets, future work would be to limit the number
we track with an LRU-eviction cache.

PiperOrigin-RevId: 351875768