ScrollablePositionedList: Replace front/back with primary/secondary
Instead of having logic everywhere to determine whether the front
list or back list is the currently active one, have designated
"primary" and "secondary" lists, and make the secondary list the
new primary one as necessary when scrolling completes.

(I alternatively considered naming them "active" and "transitionary",
which better describes their usage.  However, one property of
"primary" and "secondary" that I prefer is that they are more clearly
tied together.)

The primary list will always be the below the secondary list.  I'm
putting the primary on bottom so that in the normal case (where the
secondary list is not visible), we eventually can remove the
`FadeTransition` widget for the primary list.

PiperOrigin-RevId: 354670574