ScrollablePositionedList: Fix a scrollTo case with a non-zero alignment
If the list is manually scrolled, `scrollTo` with a non-zero
alignment could scroll to the wrong list item.  I noticed that
`_startScroll` set the target item but not the alignment for the
secondary (nee front) list view, which is inconsistent.  This
consequently then would cause the `PositionedListView` to be aligned
to the wrong item when rebuilt.

Possibly fixes https://github.com/google/flutter.widgets/issues/117.

PiperOrigin-RevId: 354670592